requirements.txt
torch~=1.13.1
onnxruntime~=1.14.1
torchvision~=0.14.1
numpy~=1.24.2
Pillow~=9.4.0
banana-dev